1.更新部分代码警告、过长换行等规范处理
Former-commit-id: 554df96e4e2c3a160b56546fcef81d2ed9eb047c
Former-commit-id: 908b34b2431eb6daa0736b2d1bfefc6cdfe9464d

+++ lib/DevApp/src/main/java/dev/utils/common/assist/Averager.java
@@ -46,9 +46,9 @@ public class Averager {
         if (mNumLists.size() == 0) {
             return 0;
         } else {
-            Float sum = 0f;
+            float sum = 0f;
             for (int i = 0, len = mNumLists.size(); i < len; i++) {
-                sum = sum.floatValue() + mNumLists.get(i).floatValue();
+                sum = sum + mNumLists.get(i).floatValue();
             }
             return sum / mNumLists.size();
         }
